feat: refactor code && fix bugs
