+++ client/src/components/Labels/Lebels.module.scss
@@ -1,0 +1,27 @@
+.label{
+    width: 20px;
+    padding: 0px;
+    padding-top: 4px;
+    margin-right: 5px;
+}
+.text{
+    margin-top: -4px;
+    padding-top: 2px;
+    color: #fff;
+    text-orientation: upright;
+    writing-mode: vertical-rl;
+    word-spacing: -11px;
+    font-size: 15px;
+    cursor: pointer;
+    transition: 0.3s;
+}
+.triangle{
+    transition: 0.3s;
+}
+.arrowDown {
+    width: 0;
+    height: 0;
+    border-left: 10px solid transparent;
+    border-right: 10px solid transparent;
+    border-top: 10px solid red;
+}
